// test/integration/test/ports/engines/queries/filters/field_reference/_fixture/mixed/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly TestModel: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly string: CodecTypes['pg/text@1']['output'] | null;
      readonly string2: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly int: CodecTypes['pg/int4@1']['output'] | null;
      readonly int2: ReadonlyArray<CodecTypes['pg/int4@1']['output']>;
      readonly bInt: CodecTypes['pg/int8@1']['output'] | null;
      readonly bInt2: ReadonlyArray<CodecTypes['pg/int8@1']['output']>;
      readonly float: CodecTypes['pg/float8@1']['output'] | null;
      readonly float2: ReadonlyArray<CodecTypes['pg/float8@1']['output']>;
      readonly bytes: CodecTypes['pg/bytea@1']['output'] | null;
      readonly bytes2: ReadonlyArray<CodecTypes['pg/bytea@1']['output']>;
      readonly bool: CodecTypes['pg/bool@1']['output'] | null;
      readonly bool2: ReadonlyArray<CodecTypes['pg/bool@1']['output']>;
      readonly dt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly dt2: ReadonlyArray<CodecTypes['pg/timestamptz@1']['output']>;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly TestModel: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly string: CodecTypes['pg/text@1']['input'] | null;
      readonly string2: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly int: CodecTypes['pg/int4@1']['input'] | null;
      readonly int2: ReadonlyArray<CodecTypes['pg/int4@1']['input']>;
      readonly bInt: CodecTypes['pg/int8@1']['input'] | null;
      readonly bInt2: ReadonlyArray<CodecTypes['pg/int8@1']['input']>;
      readonly float: CodecTypes['pg/float8@1']['input'] | null;
      readonly float2: ReadonlyArray<CodecTypes['pg/float8@1']['input']>;
      readonly bytes: CodecTypes['pg/bytea@1']['input'] | null;
      readonly bytes2: ReadonlyArray<CodecTypes['pg/bytea@1']['input']>;
      readonly bool: CodecTypes['pg/bool@1']['input'] | null;
      readonly bool2: ReadonlyArray<CodecTypes['pg/bool@1']['input']>;
      readonly dt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly dt2: ReadonlyArray<CodecTypes['pg/timestamptz@1']['input']>;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly testModel: {
      readonly bInt: CodecTypes['pg/int8@1']['output'] | null;
      readonly bInt2: ReadonlyArray<CodecTypes['pg/int8@1']['output']>;
      readonly bool: CodecTypes['pg/bool@1']['output'] | null;
      readonly bool2: ReadonlyArray<CodecTypes['pg/bool@1']['output']>;
      readonly bytes: CodecTypes['pg/bytea@1']['output'] | null;
      readonly bytes2: ReadonlyArray<CodecTypes['pg/bytea@1']['output']>;
      readonly dt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly dt2: ReadonlyArray<CodecTypes['pg/timestamptz@1']['output']>;
      readonly float: CodecTypes['pg/float8@1']['output'] | null;
      readonly float2: ReadonlyArray<CodecTypes['pg/float8@1']['output']>;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly int: CodecTypes['pg/int4@1']['output'] | null;
      readonly int2: ReadonlyArray<CodecTypes['pg/int4@1']['output']>;
      readonly string: CodecTypes['pg/text@1']['output'] | null;
      readonly string2: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly testModel: {
      readonly bInt: CodecTypes['pg/int8@1']['input'] | null;
      readonly bInt2: ReadonlyArray<CodecTypes['pg/int8@1']['input']>;
      readonly bool: CodecTypes['pg/bool@1']['input'] | null;
      readonly bool2: ReadonlyArray<CodecTypes['pg/bool@1']['input']>;
      readonly bytes: CodecTypes['pg/bytea@1']['input'] | null;
      readonly bytes2: ReadonlyArray<CodecTypes['pg/bytea@1']['input']>;
      readonly dt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly dt2: ReadonlyArray<CodecTypes['pg/timestamptz@1']['input']>;
      readonly float: CodecTypes['pg/float8@1']['input'] | null;
      readonly float2: ReadonlyArray<CodecTypes['pg/float8@1']['input']>;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly int: CodecTypes['pg/int4@1']['input'] | null;
      readonly int2: ReadonlyArray<CodecTypes['pg/int4@1']['input']>;
      readonly string: CodecTypes['pg/text@1']['input'] | null;
      readonly string2: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;